Fix floor function to use math.floor
Updated the implementation of the `floor` function to use `math.floor` from the math module. Corrected the return type annotation to `int`, as `math.floor` returns an integer value.

+++ 0x00-python_variable_annotations/2-floor.py
@@ -1,4 +1,5 @@
 #!/usr/bin/env python3
+
 import math
 """
 function floor which takes a float n as argument
@@ -6,7 +7,7 @@ returns the floor of the float.
 """
 
 
-def floor(n: float) -> float:
+def floor(n: float) -> int:
     """
     returns the floor of the float.
     """
